requirements.txt
beautifulsoup4
cloudscraper
fake-useragent
flask
lxml
playwright
ffmpeg-python
python-dotenv
requests
tqdm
yt-dlp